+++ js/workflows/pkm-retrieval-config/return_scoring_config_v1.js
@@ -1,12 +1,13 @@
 'use strict';
 
-const { CONFIG_V1 } = require('../../../src/server/config.js');
+const { getConfig } = require('../../../src/server/config.js');
 
 module.exports = async function ({ $json }) {
+  const config = await getConfig();
   return [{
     json: {
       ...$json,
-      config: CONFIG_V1,
+      config,
     },
   }];
 };